Проверка корректности хода
